Revert "feat: profile concurrent transactions (#2105)"
This reverts commit e43ce746ccbe0b1431e57afa01812beb2fe8cead.
