+++ Sources/Sentry/SentrySpan.m
@@ -2,10 +2,7 @@
 #import "NSDate+SentryExtras.h"
 #import "NSDictionary+SentrySanitize.h"
 #import "SentryCurrentDate.h"
-#import "SentryLog.h"
 #import "SentryNoOpSpan.h"
-#import "SentrySpanId.h"
-#import "SentryTime.h"
 #import "SentryTraceHeader.h"
 #import "SentryTracer.h"
 
@@ -24,8 +21,6 @@ SentrySpan ()
 - (instancetype)initWithTracer:(SentryTracer *)tracer context:(SentrySpanContext *)context
 {
     if (self = [super init]) {
-        SENTRY_LOG_DEBUG(
-            @"Starting span %@ with tracer %@", context.spanId.sentrySpanIdString, tracer);
         _tracer = tracer;
         _context = context;
         self.startTimestamp = [SentryCurrentDate date];
@@ -45,7 +40,6 @@ SentrySpan ()
                               description:(nullable NSString *)description
 {
     if (self.tracer == nil) {
-        SENTRY_LOG_DEBUG(@"No tracer, returning no-op span");
         return [SentryNoOpSpan shared];
     }
