+++ Sources/Sentry/SentryTracer.m
@@ -15,7 +15,6 @@
 #import "SentrySpan.h"
 #import "SentrySpanContext.h"
 #import "SentrySpanId.h"
-#import "SentryTime.h"
 #import "SentryTraceContext.h"
 #import "SentryTransaction+Private.h"
 #import "SentryTransaction.h"
@@ -45,12 +44,14 @@ SentryTracer ()
 @property (nonatomic) BOOL isWaitingForChildren;
 @property (nonatomic) NSTimeInterval idleTimeout;
 @property (nonatomic, nullable, strong) SentryDispatchQueueWrapper *dispatchQueueWrapper;
+@property (nonatomic, assign, readwrite) BOOL isProfiling;
 
 @end
 
 @implementation SentryTracer {
     BOOL _waitForChildren;
     SentryTraceContext *_traceContext;
+    SentryProfilesSamplerDecision *_profilesSamplerDecision;
     NSMutableDictionary<NSString *, id> *_tags;
     NSMutableDictionary<NSString *, id> *_data;
     dispatch_block_t _idleTimeoutBlock;
@@ -68,11 +69,19 @@ SentryTracer ()
 static NSObject *appStartMeasurementLock;
 static BOOL appStartMeasurementRead;
 
+#if SENTRY_TARGET_PROFILING_SUPPORTED
+static SentryProfiler *_Nullable profiler;
+static NSLock *profilerLock;
+#endif
+
 + (void)initialize
 {
     if (self == [SentryTracer class]) {
         appStartMeasurementLock = [[NSObject alloc] init];
         appStartMeasurementRead = NO;
+#if SENTRY_TARGET_PROFILING_SUPPORTED
+        profilerLock = [[NSLock alloc] init];
+#endif
     }
 }
 
@@ -140,6 +149,7 @@ static BOOL appStartMeasurementRead;
         _children = [[NSMutableArray alloc] init];
         self.hub = hub;
         self.isWaitingForChildren = NO;
+        _profilesSamplerDecision = profilesSamplerDecision;
         _waitForChildren = waitForChildren;
         _tags = [[NSMutableDictionary alloc] init];
         _data = [[NSMutableDictionary alloc] init];
@@ -164,10 +174,19 @@ static BOOL appStartMeasurementRead;
             initFrozenFrames = currentFrames.frozen;
         }
 #endif // SENTRY_HAS_UIKIT
-
 #if SENTRY_TARGET_PROFILING_SUPPORTED
-        if (profilesSamplerDecision.decision == kSentrySampleDecisionYes) {
-            [SentryProfiler startForSpanID:transactionContext.spanId hub:hub];
+        if (_profilesSamplerDecision.decision == kSentrySampleDecisionYes) {
+            [profilerLock lock];
+            if (profiler == nil) {
+                profiler = [[SentryProfiler alloc] init];
+                SENTRY_LOG_DEBUG(@"Starting profiler.");
+#    if SENTRY_HAS_UIKIT
+                framesTracker.currentTracer = self;
+                [framesTracker resetProfilingTimestamps];
+#    endif // SENTRY_HAS_UIKIT
+                [profiler start];
+            }
+            [profilerLock unlock];
         }
 #endif // SENTRY_TARGET_PROFILING_SUPPORTED
     }
@@ -245,7 +264,6 @@ static BOOL appStartMeasurementRead;
                                            sampled:_rootSpan.context.sampled];
     context.spanDescription = description;
 
-    SENTRY_LOG_DEBUG(@"Starting child span under %@", parentId.sentrySpanIdString);
     SentrySpan *child = [[SentrySpan alloc] initWithTracer:self context:context];
     @synchronized(_children) {
         [_children addObject:child];
@@ -429,9 +447,23 @@ static BOOL appStartMeasurementRead;
         self.finishCallback = nil;
     }
 
-    if (_hub == nil) {
+    if (_hub == nil)
         return;
+
+#if SENTRY_TARGET_PROFILING_SUPPORTED
+    SentryScreenFrames *frameInfo;
+    if (_profilesSamplerDecision.decision == kSentrySampleDecisionYes) {
+        SENTRY_LOG_DEBUG(@"Stopping profiler.");
+        [profilerLock lock];
+        [profiler stop];
+#    if SENTRY_HAS_UIKIT
+        frameInfo = SentryFramesTracker.sharedInstance.currentFrames;
+        [SentryFramesTracker.sharedInstance resetProfilingTimestamps];
+        SentryFramesTracker.sharedInstance.currentTracer = nil;
+#    endif // SENTRY_HAS_UIKIT
+        [profilerLock unlock];
     }
+#endif // SENTRY_TARGET_PROFILING_SUPPORTED
 
     [_hub.scope useSpan:^(id<SentrySpan> _Nullable span) {
         if (span == self) {
@@ -459,10 +491,6 @@ static BOOL appStartMeasurementRead;
         }
     }
 
-#if SENTRY_TARGET_PROFILING_SUPPORTED
-    [SentryProfiler stopProfilingSpan:self.rootSpan];
-#endif // SENTRY_TARGET_PROFILING_SUPPORTED
-
     SentryTransaction *transaction = [self toTransaction];
 
     // Prewarming can execute code up to viewDidLoad of a UIViewController, and keep the app in the
@@ -474,16 +502,30 @@ static BOOL appStartMeasurementRead;
         SENTRY_LOG_INFO(@"Auto generated transaction exceeded the max duration of %f seconds. Not "
                         @"capturing transaction.",
             SENTRY_AUTO_TRANSACTION_MAX_DURATION);
-#if SENTRY_TARGET_PROFILING_SUPPORTED
-        [SentryProfiler dropTransaction:transaction];
-#endif // SENTRY_TARGET_PROFILING_SUPPORTED
         return;
     }
-    [_hub captureTransaction:transaction withScope:_hub.scope];
+
+    NSMutableArray<SentryEnvelopeItem *> *additionalEnvelopeItems = [NSMutableArray array];
 
 #if SENTRY_TARGET_PROFILING_SUPPORTED
-    [SentryProfiler linkTransaction:transaction];
-#endif // SENTRY_TARGET_PROFILING_SUPPORTED
+    if (_profilesSamplerDecision.decision == kSentrySampleDecisionYes) {
+        [profilerLock lock];
+        if (profiler != nil) {
+            SentryEnvelopeItem *profile = [profiler buildEnvelopeItemForTransaction:transaction
+                                                                                hub:_hub
+                                                                          frameInfo:frameInfo];
+            if (profile != nil) {
+                [additionalEnvelopeItems addObject:profile];
+            }
+            profiler = nil;
+        }
+        [profilerLock unlock];
+    }
+#endif
+
+    [_hub captureTransaction:transaction
+                      withScope:_hub.scope
+        additionalEnvelopeItems:additionalEnvelopeItems];
 }
 
 - (void)trimEndTimestamp
@@ -745,6 +787,16 @@ static BOOL appStartMeasurementRead;
     }
     return nil;
 }
+
+#if SENTRY_TARGET_PROFILING_SUPPORTED
+- (BOOL)isProfiling
+{
+    [profilerLock lock];
+    BOOL isRunning = profiler.isRunning;
+    [profilerLock unlock];
+    return isRunning;
+}
+#endif // SENTRY_TARGET_PROFILING_SUPPORTED
 
 @end
