+++ Tests/SentryTests/Performance/SentryTracerTests.swift
@@ -728,6 +728,42 @@ class SentryTracerTests: XCTestCase {
         XCTAssertEqual(dict, [fixture.testKey: fixture.testValue])
     }
     
+#if os(iOS) || os(macOS) || targetEnvironment(macCatalyst)
+    func testCapturesProfile_whenProfilingEnabled() {
+        let scope = Scope()
+        let options = Options()
+        options.profilesSampleRate = 1.0
+        options.tracesSampleRate = 1.0
+        let client = TestClient(options: options)!
+        let hub = TestHub(client: client, andScope: scope)
+        
+        let tracer = hub.startTransaction(transactionContext: fixture.transactionContext) as! SentryTracer
+        tracer.finish()
+        hub.group.wait()
+        
+        XCTAssertEqual("profile", hub.capturedEventsWithScopes.first?.additionalEnvelopeItems.first?.header.type)
+    }
+    
+    func testDoesNotCapturesProfile_whenProfilingDisabled() {
+        let scope = Scope()
+        let options = Options()
+        options.profilesSampleRate = 0.0
+        options.tracesSampleRate = 1.0
+        let client = TestClient(options: options)!
+        let hub = TestHub(client: client, andScope: scope)
+        
+        let tracer = hub.startTransaction(transactionContext: fixture.transactionContext) as! SentryTracer
+        tracer.finish()
+        hub.group.wait()
+        
+        if let items = hub.capturedEventsWithScopes.first?.additionalEnvelopeItems {
+            for item in items {
+                XCTAssertNotEqual("profile", item.header.type)
+            }
+        }
+    }
+#endif
+    
     private func advanceTime(bySeconds: TimeInterval) {
         fixture.currentDateProvider.setDate(date: fixture.currentDateProvider.date().addingTimeInterval(bySeconds))
